Fetching of activity PoC
Buuuut it turns out Strava's authorization for these things has changed.
You cannot get the details of an activity any more without express
authorization of the user in question. In other words, we cannot use it
as it currently stands. Committing it as a sort of reminder of how to do
this. Guess I will have to change it to instead parse the actual HTML
page... Blergh, just when I got this looking pretty.