Redo structure so plugins share a Handler trait
Think I might also need a mutable handler for future purposes. Possibly
current ones as well: the calc plugin uses a core that it gets from the
disk. The core needs to be mutable, so the current Box approach might
not do. At the very least in the future for a football plugin, there
will be the need to cache games so the struct doing the Handler'ing will
need to be able to mute its state.