strava: different 0-width char for leaderboard
The zero-width space was giving issues on iterm2 and other macos
terminals which for some reason decide to render it as taking up one
character. This ruins weechat interface and similar. Using this other
character seems to do the same trick and does not fuck up the interface.
Hope it works the same for everyone.