Bump reqwest to =0.9.10
Needed to upgrade past 0.8.x as building on ubuntu did not work any
more: mismatched openssl versions. The 0.9.x line fixes that.

They are already at 0.9.20, but that gave futures issues in my code.
Would have to write everything with async stuff and what not. Could not
be bothered to learn the details about this right now (will I be forced
into a callbacklike js hell again?), found out 0.9.10 did not give me
this issue yet so upgraded to that.