Add an AsyncMutableHandler
Rust does not allow traits to have async functions. There is a crate to
work around that, presumably it scrubs some information before it
reaches the compiler? I do not understand exactly. Either way, add that
crate and a new type for those that need to be async and mutable.