Bump own version
New type of handler trait deserves a version bump I reckon. Now to still
make full use of these being async (and adapt it to the others?) in
order to just fire off all of the handlers, kill after x seconds if
needed, and never stop the following messages to be delayed.