Add the "default" logging setup
Basically the one is an interface that a bunch of libraries seem to use.
The other is an implementation that handles the actual printing etc.
This one seemed to be used and simple enough. Needed it as I was
debugging reqwest and figured I should probably eventually switch my
crude println! stuff to it as well.